backend/requirements.txt
# MIRIP Backend Dependencies
# Version: SPEC-BACKEND-001

# ======================
# Core Framework
# ======================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ======================
# Cache & Storage
# ======================
redis>=5.0.0

# ======================
# Utilities
# ======================
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.25.0

# ======================
# Logging
# ======================
structlog>=23.2.0

# ======================
# ML Dependencies (SPEC-AI-001)
# DINOv2 Baseline AI Evaluation Model
# ======================
torch>=2.1.0
transformers>=4.35.0
timm>=0.9.0
pytorch-lightning>=2.1.0
wandb>=0.16.0
scikit-learn>=1.3.0

# ======================
# Image Processing (SPEC-AI-001)
# ======================
opencv-python>=4.8.0
pillow>=10.0.0
albumentations>=1.3.0

# ======================
# OpenAI (Feedback Generation)
# ======================
# openai>=1.0.0
